chore: 주석 추가 및 코드 정리
설명을 위한 주석을 추가했고, 반복되는 코드나 불필요한 코드를 정리했다.
