+++ src/utils/search.ts
@@ -1,3 +1,5 @@
+// 한글 초성 검색을 가능하게 하는 기능, 그냥 귀찮아서 구글링해서 찾음
+
 const CHO_HANGUL = [
   'ㄱ',
   'ㄲ',
